Removed updateday and storylineusedate.
For the first of my executive decisions to cut cruft, updateday and
storylineusedate are no more.  updateday is redundant with a properly-
configured tzoffset, and storylineusedate was a workaround for an older
storyline file format I'm not going to use for the next version.

Sanity checking for updatetime is also now in play.